Adding inheritance to Backend class
- Preparing for JSON->kll and kll->JSON compilation modes
